Move toolbox to become the main topsail folder
This commit renames the current toolbox folder to be
the main topsail folder for all the cli options.
