+++ topsail/benchmarking.py
@@ -1,4 +1,4 @@
-from toolbox._common import RunAnsibleRole, AnsibleRole, AnsibleMappedParams
+from topsail._common import RunAnsibleRole, AnsibleRole, AnsibleMappedParams
 
 
 class Benchmarking:
