+++ topsail/entitlement.py
@@ -1,4 +1,4 @@
-from toolbox._common import RunAnsibleRole, AnsibleRole
+from topsail._common import RunAnsibleRole, AnsibleRole
 
 
 class Entitlement:
