+++ topsail/gpu-operator/_helm_deploy_operator.sh
