+++ topsail/gpu-operator/diagnose.sh
